src/kerastb/requirements.txt
tensorflow==2.7.0
numpy==1.21.5
opencv-python==4.5.4.58
matplotlib==3.4.3
azure-functions
azure-storage-blob
Pillow